LPS-75890 Refactor how middlewares are linked to LCSClient
