+++ analytics-client-js/src/LCSClient/middlewares/defaults.js
@@ -1,0 +1,15 @@
+import bootstrap from './bootstrap';
+import meta from './meta';
+import openGraph from './opengraph';
+
+export {
+	bootstrap,
+	meta,
+	openGraph,
+};
+
+export default [
+	bootstrap,
+	meta,
+	openGraph
+];
