--- analytics-client-js/src/middlewares/defaults.js
@@ -1,3 +1,0 @@
-import './boot';
-import './meta';
-import './opengraph';
